style: format with forge

+++ src/test/Token.sol
@@ -5,11 +5,11 @@ import "@openzeppelin/contracts/access/Ownable.sol";
 import "@openzeppelin/contracts/token/ERC20/ERC20.sol";
 
 contract Token is ERC20, Ownable {
-    constructor(uint initialSupply) ERC20("Token", "TOKE") {
+    constructor(uint256 initialSupply) ERC20("Token", "TOKE") {
         _mint(owner(), initialSupply);
     }
 
-    function mint(uint amount) external onlyOwner {
+    function mint(uint256 amount) external onlyOwner {
         _mint(owner(), amount);
     }
 }
